Fix when unchecking button from category settings
Taking a look to the values of every variable, for instance, under the function canTopicBeMarkedAsCancelled, I realized that the return statement was returning a misleading value.

Context:
- Enable the plugin and enable Sold and Cancelled buttons under some random category.
- Then try to uncheck Cancelled button from that category. You will see that the button Cancelled still appears.

The error comes from extend-topic-for-sold-button.js.es6, at the time of defining the variable ```const enable_cancelled_button``` (it affects to all of the buttons, but I will focus on this). The thing is that at the time of executing the return statement, all '&&' returns true, because enable_cancelled_button is a string (which its value is "false").